Document action parsing router helper and move private helper functions into /internal
